+++ mayan/apps/templating/classes.py
@@ -62,6 +62,7 @@ class Template:
                         'builtins': [
                             'mathfilters.templatetags.mathfilters',
                             'mayan.apps.templating.templatetags.templating_datetime_tags',
+                            'mayan.apps.templating.templatetags.templating_json_tags',
                             'mayan.apps.templating.templatetags.templating_regex_tags',
                             'mayan.apps.templating.templatetags.templating_tags'
                         ]